feat(Toolbar): Break toolbar out as its own component
Allows the user to select whether they would like the simple collapse icon or
the material-ui theme toolbar.
